feat: shift non-e2e tests into sandbox_simple_test.go
